miniproject00/pyproject.toml
[project]
name = "mini-project-00"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cohere>=5.20.1",
    "google-genai>=1.56.0",
    "groq>=1.0.0",
    "mistralai>=1.10.0",
    "openai>=2.14.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.1",
]
